term: fix deleted line position of deleteLines
The original implementation in term.coffee causes problems for programs such as `tmux` and `vim`, due to the fact that it inserted the blank line first. After inserting the blank line, the `@y` position is no longer the current position, but actually on the previous line. This causes the previous line to be deleted instead.

This patch simply reverses the order of the deletion and the insertion.