linkify: fix parsing for nested links
Sometimes a link can contain a substring that may be parsed as another type of link, e.g. http://example.com/something@example.com. In this case, the original implementation will cause errors because the link tags are nested. This can be fixed by matching them first and later replace them with the correct linkified tags.